django/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.2
psycopg2-binary==2.9.10
django-cors-headers==4.6.0